Clean up assistant controllers
